poioson[fix] pep8 and lambtha as a float

+++ math/0x03-probability/poisson.py
@@ -14,4 +14,4 @@ class Poisson:
                 raise TypeError('data must be a list')
             if len(data) <= 1:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = sum(data) / len(data)
+            self.lambtha = float(sum(data) / len(data))
